1.21.1: match upstream code style
The port is meant to go upstream as pull requests, so its additions now follow
docs/code-style.md (in CoFHCore), measured from upstream:
- Porting comments are gone: version tags, migration narration, doc pointers.
  A non-obvious reason keeps one short line.
- Imports use IntelliJ group order, unused ones are removed, and inline
  qualified names became imports.
- Upstream forms the port changed without needing to are restored: commented-out
  code, parameter names, line wrapping, member order, redundant casts.

No behaviour change. All four repos build, and TE/TD boot headless.

+++ src/main/java/cofh/core/client/particle/SpriteParticle.java
@@ -108,10 +108,10 @@ public abstract class SpriteParticle extends ColorParticle {
         float v0 = sprite.getV0();
         float v1 = sprite.getV1();
 
-        consumer.addVertex((float) (x + a), (float) (y + b), (float) (z)).setUv(u1, v0).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
-        consumer.addVertex((float) (x - b), (float) (y + a), (float) (z)).setUv(u0, v0).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
-        consumer.addVertex((float) (x - a), (float) (y - b), (float) (z)).setUv(u0, v1).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
-        consumer.addVertex((float) (x + b), (float) (y - a), (float) (z)).setUv(u1, v1).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
+        consumer.addVertex(x + a, y + b, z).setUv(u1, v0).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
+        consumer.addVertex(x - b, y + a, z).setUv(u0, v0).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
+        consumer.addVertex(x - a, y - b, z).setUv(u0, v1).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
+        consumer.addVertex(x + b, y - a, z).setUv(u1, v1).setColor(c0.r, c0.g, c0.b, c0.a).setLight(packedLight);
     }
 
 }
